Move Challenge -> com.unifina.service
