+++ grails-app/controllers/com/unifina/controller/LoginApiController.groovy
@@ -4,7 +4,6 @@ package com.unifina.controller
 import com.unifina.domain.SignupMethod
 import com.unifina.domain.User
 import com.unifina.domain.Userish
-import com.unifina.security.Challenge
 import com.unifina.service.*
 import grails.converters.JSON
