+++ src/groovy/com/unifina/service/Challenge.groovy
@@ -1,4 +1,4 @@
-package com.unifina.security
+package com.unifina.service
 
 import com.unifina.utils.AlphanumericStringGenerator
 import org.joda.time.DateTime
