+++ test/unit/com/unifina/controller/LoginApiControllerSpec.groovy
@@ -3,7 +3,6 @@ package com.unifina.controller
 import com.unifina.domain.Key
 import com.unifina.domain.SignupMethod
 import com.unifina.domain.User
-import com.unifina.security.Challenge
 import com.unifina.service.*
 import grails.test.mixin.Mock
 import grails.test.mixin.TestFor
