+++ test/unit/com/unifina/service/ChallengeSpec.groovy
@@ -1,5 +1,4 @@
-package com.unifina.security
-
+package com.unifina.service
 
 import spock.lang.Specification
